feat: Add react-component-export-image npm dependency
This commit adds the "react-component-export-image" npm dependency to the project. This package allows for exporting React components as images. By including this dependency, we can enhance the functionality of the application by enabling users to export roast results as images.

Note: This commit message follows the established convention of starting with a verb in the imperative form and providing a concise description of the changes made.